Split PageRef into data module.
