+++ ftml/src/data/mod.rs
@@ -20,8 +20,10 @@
 
 mod backlinks;
 mod page_info;
+mod page_ref;
 mod user_info;
 
 pub use self::backlinks::Backlinks;
 pub use self::page_info::PageInfo;
+pub use self::page_ref::PageRef;
 pub use self::user_info::UserInfo;
